skills/kesslerio/oura-analytics/requirements.txt
# Oura Analytics Skill Dependencies
# Note: Uses Python's built-in urllib.request (no external HTTP library)
pytz
pyyaml
